+++ src/main.rs
@@ -5,7 +5,7 @@ use lazy_static::lazy_static;
 use pager::Pager;
 use regex::{Captures, Regex};
 use std::io::BufRead;
-use unicode_segmentation::UnicodeSegmentation;
+use termwrap::termwrap;
 use which::which;
 
 /**
@@ -325,8 +325,8 @@ fn run<T: AsRef<str>>(
         run_(command)
     } else {
         let (stdout, stderr, code) = pipe(command);
-        let stderr = term_hard_wrap(&stderr, wrap, continuation);
-        let stdout = term_hard_wrap(&stdout, wrap, continuation);
+        let stderr = termwrap(&stderr, wrap, continuation);
+        let stdout = termwrap(&stdout, wrap, continuation);
         print!("{stderr}{stdout}");
         code
     };
@@ -347,88 +347,6 @@ fn cd(dir: &std::path::Path) {
         std::env::set_current_dir(dir)
             .unwrap_or_else(|e| exit!(103, "ERROR: Could not change directory to {dir:?}: {e}"));
     }
-}
-
-/**
-Wrap text with ANSI color codes to a fixed number of columns
-*/
-fn term_hard_wrap(s: &str, width: usize, continuation: &str) -> String {
-    if width == 0 {
-        return s.to_string();
-    }
-
-    // Input without ANSI color codes to graphemes
-    let b = String::from_utf8(strip_ansi_escapes::strip(s.as_bytes()).unwrap()).unwrap();
-    let mut gb = b.graphemes(true).collect::<Vec<_>>();
-
-    // Input graphemes
-    let mut ga = s.graphemes(true).collect::<Vec<_>>();
-
-    // Continuation graphemes
-    let gc = continuation.graphemes(true).collect::<Vec<_>>();
-    let cw = gc.len();
-    let cwp = cw + 1;
-    let w = width - cw; // initial max width; leave space for the continuation
-
-    let mut r = String::new(); // result
-    let mut l = 0; // current column
-    let mut ca; // current grapheme from `ga`
-    let mut cb; // current grapheme from `gb`
-
-    // Internal function update state
-    fn update(ca: &str, r: &mut String, l: &mut usize, counts: bool) {
-        if ca == "\t" {
-            // Tab counts as a color code for some reason, so it always counts
-            // against the line length.
-            // `8 - *l % 8` is he number of spaces from the current column to
-            // the nearest multiple of 8.
-            for _ in 0..(8 - *l % 8) {
-                *l += 1;
-                r.push(' ');
-            }
-        } else {
-            // Not a tab, so it only counts if not a color code.
-            if counts {
-                *l = if ca == "\n" { 0 } else { *l + 1 };
-            }
-            r.push_str(ca);
-        }
-    }
-
-    while !ga.is_empty() {
-        // Process line until the initial max width
-        while l < w {
-            if ga.is_empty() {
-                break;
-            }
-            ca = ga.remove(0);
-            cb = gb.remove(0);
-
-            // Process extra ANSI color code graphemes
-            while ca != cb && !ga.is_empty() {
-                update(ca, &mut r, &mut l, false);
-                ca = ga.remove(0);
-            }
-
-            // Process regular graphemes
-            update(ca, &mut r, &mut l, true);
-        }
-
-        // Process the last character(s) of the line
-        let gal = ga.len();
-        if gal == cw || (gal == cwp && ga[1] == "\n") {
-            // Continuation width characters left in the input or the current line, so just push
-            // them
-            r.push_str(ga.remove(0));
-        } else if gal != 0 {
-            // Unless done, insert the continuation
-            r.push_str(continuation);
-            r.push('\n');
-        }
-        l = 0;
-    }
-
-    r
 }
 
 /**